refactor: replace "/setup" with "/videos/new"
