+++ app/__e2e__/basic.test.ts
@@ -1,6 +1,6 @@
 import { expect, test } from "@playwright/test";
 
-test("index => setup", async ({ page }) => {
+test("/ => /videos/new", async ({ page }) => {
   await page.goto("/");
   await expect(page).toHaveTitle("ytsub-v3");
 
@@ -10,12 +10,12 @@ test("index => setup", async ({ page }) => {
     .fill("EnPYXckiUVg");
   await page.locator("data-test=search-form >> input >> nth=0").press("Enter");
 
-  // Navigate to /setup
-  await expect(page).toHaveURL("/setup?videoId=EnPYXckiUVg");
+  // Navigate to /vides/new
+  await expect(page).toHaveURL("/videos/new?videoId=EnPYXckiUVg");
 });
 
-test("setup => videos/id", async ({ page }) => {
-  await page.goto("/setup?videoId=EnPYXckiUVg");
+test("/videos/new => videos/id", async ({ page }) => {
+  await page.goto("/videos/new?videoId=EnPYXckiUVg");
 
   // Check readonly input
   await expect(
