+++ app/misc/routes.ts
@@ -6,9 +6,7 @@
 
 export const R = {
   "/": "/",
-  "/setup": "/setup",
   "/videos/new": "/videos/new",
-  "/videos/create": "/videos/create",
   "/videos/$id": (id: number) => `/videos/${id}`,
   "/users/me": "/users/me",
   "/users/register": "/users/register",
